# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ChatterPay Global API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterBody {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
  phone?: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export type CurrentUserRole =
  (typeof CurrentUserRole)[keyof typeof CurrentUserRole];

export const CurrentUserRole = {
  user: "user",
  foreign_client: "foreign_client",
  admin: "admin",
} as const;

export interface CurrentUser {
  id: number;
  username: string;
  email?: string;
  role: CurrentUserRole;
  kshBalance: number;
  usdtBalance: number;
  isBanned: boolean;
}

export interface AuthResponse {
  token: string;
  user: CurrentUser;
}

export interface User {
  id: number;
  username: string;
  email?: string;
  phone?: string;
  kshBalance: number;
  usdtBalance: number;
  isBanned: boolean;
  createdAt: string;
  assignedClientsCount?: number;
  hasActiveContract?: boolean;
}

export interface BanUserBody {
  banned: boolean;
}

export interface ForeignClient {
  id: number;
  username: string;
  usdtBalance: number;
  country?: string;
  isOnline: boolean;
  createdAt: string;
  assignedUsersCount?: number;
}

export interface CreateForeignClientBody {
  /** @minLength 3 */
  username: string;
  /** @minLength 6 */
  password: string;
  usdtBalance?: number;
  country?: string;
}

export type ContractPlan = (typeof ContractPlan)[keyof typeof ContractPlan];

export const ContractPlan = {
  "24h": "24h",
} as const;

export interface Contract {
  id: number;
  userId: number;
  plan: ContractPlan;
  costKsh: number;
  maxClients: number;
  isActive: boolean;
  expiresAt: string;
  createdAt: string;
}

export type CreateContractBodyPlan =
  (typeof CreateContractBodyPlan)[keyof typeof CreateContractBodyPlan];

export const CreateContractBodyPlan = {
  "24h": "24h",
} as const;

export interface CreateContractBody {
  plan: CreateContractBodyPlan;
}

export interface Wallet {
  kshBalance: number;
  usdtBalance: number;
  userId: number;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  deposit: "deposit",
  withdrawal: "withdrawal",
  contract_purchase: "contract_purchase",
  earning: "earning",
} as const;

export type TransactionCurrency =
  (typeof TransactionCurrency)[keyof typeof TransactionCurrency];

export const TransactionCurrency = {
  KSH: "KSH",
  USDT: "USDT",
} as const;

export interface Transaction {
  id: number;
  userId: number;
  type: TransactionType;
  amount: number;
  currency: TransactionCurrency;
  description?: string;
  createdAt: string;
}

export interface MpesaDepositBody {
  phone: string;
  /** @minimum 1 */
  amount: number;
}

export interface MpesaDepositResponse {
  checkoutRequestId?: string;
  merchantRequestId?: string;
  responseDescription?: string;
  message: string;
}

export interface Assignment {
  id: number;
  userId: number;
  foreignClientId: number;
  assignedAt: string;
  user?: User;
  foreignClient?: ForeignClient;
  chatRoomId?: number;
}

export interface CreateAssignmentBody {
  userId: number;
  foreignClientId: number;
}

export type ChatRoomParticipant = {
  id: number;
  username: string;
  isOnline: boolean;
};

export interface ChatRoom {
  id: number;
  assignmentId: number;
  lastMessage?: string;
  lastMessageAt?: string;
  unreadCount: number;
  participant: ChatRoomParticipant;
}

export type MessageSenderRole =
  (typeof MessageSenderRole)[keyof typeof MessageSenderRole];

export const MessageSenderRole = {
  user: "user",
  foreign_client: "foreign_client",
} as const;

export interface Message {
  id: number;
  roomId: number;
  senderId: number;
  senderRole: MessageSenderRole;
  senderUsername: string;
  content: string;
  createdAt: string;
}

export interface SendMessageBody {
  /** @minLength 1 */
  content: string;
}

export interface AdminDashboard {
  totalUsers: number;
  totalForeignClients: number;
  activeContracts: number;
  totalDepositsKsh: number;
  totalEarningsUsdt: number;
  pendingWithdrawals: number;
  totalMessages: number;
  recentUsers: User[];
}

export type WithdrawalCurrency =
  (typeof WithdrawalCurrency)[keyof typeof WithdrawalCurrency];

export const WithdrawalCurrency = {
  KSH: "KSH",
  USDT: "USDT",
} as const;

export type WithdrawalStatus =
  (typeof WithdrawalStatus)[keyof typeof WithdrawalStatus];

export const WithdrawalStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface Withdrawal {
  id: number;
  userId: number;
  amount: number;
  currency: WithdrawalCurrency;
  phone?: string;
  status: WithdrawalStatus;
  createdAt: string;
  username?: string;
}

export type WithdrawalBodyCurrency =
  (typeof WithdrawalBodyCurrency)[keyof typeof WithdrawalBodyCurrency];

export const WithdrawalBodyCurrency = {
  KSH: "KSH",
  USDT: "USDT",
} as const;

export interface WithdrawalBody {
  /** @minimum 1 */
  amount: number;
  currency: WithdrawalBodyCurrency;
  phone?: string;
}

export type NotificationType =
  (typeof NotificationType)[keyof typeof NotificationType];

export const NotificationType = {
  payment_success: "payment_success",
  contract_activated: "contract_activated",
  client_assigned: "client_assigned",
  withdrawal_approved: "withdrawal_approved",
  withdrawal_rejected: "withdrawal_rejected",
} as const;

export interface Notification {
  id: number;
  userId: number;
  type: NotificationType;
  message: string;
  isRead: boolean;
  createdAt: string;
}

export type MpesaCallbackBody = { [key: string]: unknown };