# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ChatterPay Global API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new Kenyan user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login (users, foreign clients, admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /assignments:
    get:
      operationId: listAssignments
      tags: [assignments]
      summary: List assignments for current user/client
      responses:
        "200":
          description: List of assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Assignment"
    post:
      operationId: createAssignment
      tags: [assignments]
      summary: Assign a foreign client to a Kenyan user (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAssignmentBody"
      responses:
        "201":
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Assignment"
        "400":
          description: Max clients reached or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /assignments/{assignmentId}:
    delete:
      operationId: deleteAssignment
      tags: [assignments]
      summary: Remove an assignment (admin only)
      parameters:
        - name: assignmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Assignment removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /chat/rooms/{roomId}/messages:
    get:
      operationId: listMessages
      tags: [chat]
      summary: List messages in a chat room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages in room
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message in a chat room
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        "403":
          description: No active contract or not assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    RegisterBody:
      type: object
      properties:
        username:
          type: string
          minLength: 3
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        phone:
          type: string
      required: [username, email, password]
    LoginBody:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required: [username, password]
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/CurrentUser"
      required: [token, user]
    CurrentUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [user, foreign_client, admin]
        kshBalance:
          type: number
        usdtBalance:
          type: number
        isBanned:
          type: boolean
      required: [id, username, role, kshBalance, usdtBalance, isBanned]
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        phone:
          type: string
        kshBalance:
          type: number
        usdtBalance:
          type: number
        isBanned:
          type: boolean
        createdAt:
          type: string
          format: date-time
        assignedClientsCount:
          type: integer
        hasActiveContract:
          type: boolean
      required: [id, username, kshBalance, usdtBalance, isBanned, createdAt]
...
    ForeignClient:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        usdtBalance:
          type: number
        country:
          type: string
        isOnline:
          type: boolean
        createdAt:
          type: string
          format: date-time
        assignedUsersCount:
          type: integer
      required: [id, username, usdtBalance, isOnline, createdAt]
...
    Assignment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        foreignClientId:
          type: integer
        assignedAt:
          type: string
          format: date-time
        user:
          $ref: "#/components/schemas/User"
        foreignClient:
          $ref: "#/components/schemas/ForeignClient"
        chatRoomId:
          type: integer
      required: [id, userId, foreignClientId, assignedAt]
    CreateAssignmentBody:
      type: object
      properties:
        userId:
          type: integer
        foreignClientId:
          type: integer
      required: [userId, foreignClientId]
...
    Message:
      type: object
      properties:
        id:
          type: integer
        roomId:
          type: integer
        senderId:
          type: integer
        senderRole:
          type: string
          enum: [user, foreign_client]
        senderUsername:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, roomId, senderId, senderRole, senderUsername, content, createdAt]
    SendMessageBody:
      type: object
      properties:
        content:
          type: string
          minLength: 1
      required: [content]
...