# lib/api-zod/src/generated/types/createAssignmentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ChatterPay Global API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAssignmentBody {
  userId: number;
  foreignClientId: number;
}