# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ChatterPay Global API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  email?: string;
  phone?: string;
  kshBalance: number;
  usdtBalance: number;
  isBanned: boolean;
  createdAt: Date;
  assignedClientsCount?: number;
  hasActiveContract?: boolean;
}